+++ src/DuetAPI/Commands/Generic/Code/Code.cs
@@ -323,9 +323,6 @@ namespace DuetAPI.Commands
                 KeywordType.While => "while",
                 KeywordType.Break => "break",
                 KeywordType.Continue => "continue",
-#pragma warning disable CS0618 // Type or member is obsolete
-                KeywordType.Return => "return",
-#pragma warning restore CS0618 // Type or member is obsolete
                 KeywordType.Abort => "abort",
                 KeywordType.Var => "var",
                 KeywordType.Set => "set",