+++ src/DuetControlServer/Commands/ObjectModel/PatchObjectModel.cs
@@ -20,7 +20,7 @@ namespace DuetControlServer.Commands
                 throw new InvalidOperationException("Command is only supported in non-SPI mode");
             }
 
-            if (Model.Provider.Get.UpdateFromModel(Key, Patch))
+            if (Model.Provider.Get.UpdateFromJson(Key, Patch))
             {
                 if (Model.Provider.IsUpdating && Model.Provider.Get.State.Status != MachineStatus.Updating)
                 {