+++ src/DuetControlServer/FileExecution/Job.cs
@@ -270,18 +270,9 @@ namespace DuetControlServer.FileExecution
                                     throw;
                                 }
 
+                                readCode.LogOutput = true;
                                 codes.Enqueue(readCode);
-                                codeTasks.Enqueue(
-                                    readCode
-                                        .Execute()
-                                        .ContinueWith(async task =>
-                                        {
-                                            Message result = await task;
-                                            await Logger.LogOutputAsync(result);
-                                            return result;
-                                        }, TaskContinuationOptions.RunContinuationsAsynchronously)
-                                        .Unwrap()
-                                );
+                                codeTasks.Enqueue(readCode.Execute());
                             }
                             catch (OperationCanceledException)
                             {